rename demo obfuscator modules to fit project conventions
