+++ templates/dashboard/customer/detail.html
@@ -1,5 +1,5 @@
 {% extends "dashboard/base.html" %}
-{% load demoobfuscators %}
+{% load demo_obfuscators %}
 {% load prices_i18n %}
 {% load i18n %}
 {% load materializecss %}
