+++ templates/dashboard/customer/list.html
@@ -3,7 +3,7 @@
 {% load prices_i18n %}
 {% load materializecss %}
 {% load utils %}
-{% load demoobfuscators %}
+{% load demo_obfuscators %}
 
 {% block title %}
   {% trans "Customers" context "Dashboard customers list" %} - {{ block.super }}
