+++ templates/dashboard/includes/_address.html
@@ -1,4 +1,4 @@
-{% load demoobfuscators %}
+{% load demo_obfuscators %}
 {% load format_address from i18n_address_tags %}
 {% load staticfiles %}
