+++ templates/dashboard/order/detail.html
@@ -1,6 +1,6 @@
 {% extends "dashboard/base.html" %}
 {% load babel %}
-{% load demoobfuscators %}
+{% load demo_obfuscators %}
 {% load prices_i18n %}
 {% load i18n %}
 {% load materializecss %}
